Added my project

+++ src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/NotesController.java
@@ -11,33 +11,64 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 
 @Controller
 public class NotesController {
 
-    private static UserService userService;
-    private static NoteService noteService;
+    public final UserService userService;
+    public final NoteService noteService;
 
-    public NotesController(UserService userService, NoteService noteService){
-        NotesController.userService = userService;
-        NotesController.noteService = noteService;
+    public NotesController(UserService userService, NoteService noteService) {
+        this.userService = userService;
+        this.noteService = noteService;
     }
+
+    public String createErrorMsg(List<String> errorMessages, StringBuilder message) {
+        for (int i = 0; i < errorMessages.size(); i++) {
+            if (i < (errorMessages.size() - 1)) {
+                message.append(errorMessages.get(i)).append(", Error num: ").append(i).append(": ");
+            } else {
+                message.append(errorMessages.get(i)).append(".");
+            }
+        }
+        return message.toString();
+    }
+
+    public String createSuccessMsg(List<String> successMessages, StringBuilder message) {
+        for (int i = 0; i < successMessages.size(); i++) {
+            if (i < (successMessages.size() - 1)) {
+                message.append(successMessages.get(i)).append(", And ");
+            } else {
+                message.append(successMessages.get(i)).append(".");
+            }
+        }
+        return message.toString();
+    }
+
 
     // Create New Note
     @PostMapping("/notes/add")
-    public String addNote(@ModelAttribute("NoteFormAdd") NoteFormAdd noteFormAdd, Authentication authentication, Model model) throws SQLException {
+    public String addNote(@ModelAttribute("NoteFormAdd") NoteFormAdd noteFormAdd, @ModelAttribute("noteFormEdit") NoteFormEdit noteFormEdit, Authentication authentication, Model model) throws SQLException {
+        String loggedUserName = authentication.getName();
         User loggedUser = userService.getUser(authentication.getName());
-        Note newUserNote = new Note(null,noteFormAdd.getNoteTitle(),noteFormAdd.getNoteDescription(),loggedUser.getUserId());
+        if (loggedUser == null){
+            return "login";
+        }
+        Integer loggedUserId = loggedUser.getUserId();
+
+        Note newUserNote = new Note(null, noteFormAdd.getNoteTitle(), noteFormAdd.getNoteDescription(), loggedUser.getUserId());
         int newNoteId = noteService.addNote(newUserNote, loggedUser.getUserId());
 
         if (newNoteId > 0) {
             newUserNote.setNote_id(newNoteId);
-            model.addAttribute("success_message","Successfully Added New Message");
-            model.addAttribute("newNote",newUserNote);
+            model.addAttribute("success_message", "Successfully Added New Message");
+            model.addAttribute("newNote", newUserNote);
         } else {
-            model.addAttribute("error_message","Could not add new Note");
+            model.addAttribute("error_message", "Could not add new Note");
         }
 
         List<Note> allUserNotes = noteService.getAllUserNotes(loggedUser.getUserId());
@@ -52,18 +83,23 @@ public class NotesController {
     // delete existing Note
     @PostMapping("/notes/{note_id}/delete")
     public String deleteNote(@PathVariable int note_id, Authentication authentication, Model model) throws SQLException {
+        String loggedUserName = authentication.getName();
         User loggedUser = userService.getUser(authentication.getName());
+        if (loggedUser == null){
+            return "login";
+        }
+        Integer loggedUserId = loggedUser.getUserId();
+
+
         Note getNoteUser = noteService.getNoteById(note_id);
-        System.out.println(note_id);
 
-
-        if (getNoteUser.getUser_id().equals(loggedUser.getUserId())){
+        if (getNoteUser.getUser_id().equals(loggedUser.getUserId())) {
             // apply delete
             Response newDeleteResponse = noteService.removeNote(note_id);
             model.addAttribute("delete_note_status", newDeleteResponse.isSuccess());
             model.addAttribute("logged", loggedUser.getUsername());
 
-            if (newDeleteResponse.isSuccess()){
+            if (newDeleteResponse.isSuccess()) {
                 model.addAttribute("success_message", newDeleteResponse.message);
 
             } else {
@@ -82,9 +118,31 @@ public class NotesController {
 
     }
 
-    // update Note
-    @PostMapping("/notes/{id}/update")
-    public String updateNote(Authentication authentication, Model model){
+
+    @PostMapping("/notes/{note_id}/update")
+    public String updateNote(@ModelAttribute("noteFormEdit") NoteFormEdit noteFormEdit, @PathVariable int note_id, Authentication authentication, Model model) throws SQLException {
+        String newTitle = noteFormEdit.getNoteEditTitle();
+        String newDesc = noteFormEdit.getNoteEditDescription();
+        Note noteExist = noteService.getNoteById(note_id);
+        String loggedUserName = authentication.getName();
+        User loggedUser = userService.getUser(authentication.getName());
+        if (loggedUser == null){
+            return "login";
+        }
+        Integer loggedUserId = loggedUser.getUserId();
+
+        if (noteExist == null){
+            model.addAttribute("error_message", "Note With Id" + note_id + " not found.");
+            return "result";
+        }
+        if (Objects.equals(noteExist.getUser_id(), loggedUser.getUserId())){
+            Note note = new Note(noteExist.getNote_id(), newTitle, newDesc, noteExist.getUser_id());
+            noteService.editNote(note);
+            model.addAttribute("success_message", "Things Edited");
+        } else {
+            model.addAttribute("error_message", "You are not the owner of this note you can only edit your owner of this note");
+        }
         return "result";
+
     }
 }
